Fix QT_OBJECT surface tree issue for QT6.8
