+++ ubuntu-24.10-oracular.Dockerfile
@@ -6,25 +6,24 @@ RUN apt-get -y install software-properties-common
 RUN add-apt-repository universe
 RUN apt-get update
 RUN apt-get -y install build-essential git cmake 
-RUN apt-get -y install qt6-base-dev libboost-system-dev libboost-program-options-dev 
-RUN apt-get -y install libceres-dev xtensor-dev libopencv-dev libxsimd-dev libblosc-dev libspdlog-dev 
+RUN apt-get -y install qt6-base-dev libboost-system-dev libboost-program-options-dev
+RUN apt-get -y install libceres-dev xtensor-dev libopencv-dev libxsimd-dev libblosc-dev libspdlog-dev
 RUN apt-get -y install libgsl-dev libsdl2-dev libcurl4-openssl-dev
-
 RUN apt-get -y install file
 
 COPY . /src
 RUN rm /src/CMakeCache.txt || true
 
 RUN ls /src
-RUN mkdir /build
-WORKDIR /build
+RUN mkdir /src/build
+WORKDIR /src/build
 
 RUN cmake -DVC_WITH_CUDA_SPARSE=off /src
 RUN make -j$(nproc --all)
 
 RUN cpack -G DEB -V
 
-RUN dpkg -i /build/pkgs/vc3d*.deb
+RUN dpkg -i /src/build/pkgs/vc3d*.deb
 
 RUN apt-get -y autoremove
 RUN rm -r /src
